refactor(schemas): standardize enum patterns and remove duplicate PortKindSchema
- Delete PortKindSchema.ts (duplicate of ObjectNodeTypeSchema)
- Update InputPortSchema to use ObjectNodeTypeSchema directly
- Standardize enum patterns in distillation/ (4 files)
- Standardize enum patterns in refinement/ (2 files)
- Standardize TemplateContentTypeSchema and LogLevelSchema
- All enums now use ENUM const + z.enum(ENUM) pattern
