+++ packages/schemas/src/LogLevelSchema.ts
@@ -1,4 +1,10 @@
 import { z } from "zod/v4";
 
-export const LogLevelSchema = z.enum(["info", "warn", "error", "debug"]);
+export const LOG_LEVEL_ENUM = {
+  INFO: "info",
+  WARN: "warn",
+  ERROR: "error",
+  DEBUG: "debug",
+} as const;
+export const LogLevelSchema = z.enum(LOG_LEVEL_ENUM);
 export type LogLevel = z.infer<typeof LogLevelSchema>;
