+++ packages/schemas/src/index.ts
@@ -2,8 +2,7 @@ export * from "./agent";
 export * from "./agent-runtime";
 export * from "./agent-trace";
 export * from "./ChecklistResultSchema";
-export * from "./DistillationRunSchema";
-export * from "./DistillationSchema";
+export * from "./distillation";
 export * from "./ExecutorConfigSchema";
 export * from "./ExecutorTypeSchema";
 export * from "./GithubProjectSchema";
@@ -15,10 +14,8 @@ export * from "./ObjectTypeSchema";
 export * from "./OperationConfigSchema";
 export * from "./OperationSchema";
 export * from "./OutputItemSchema";
-export * from "./PortKindSchema";
 export * from "./RecipeSchema";
-export * from "./RefinementRunSchema";
-export * from "./RefinementSchema";
+export * from "./refinement";
 export * from "./ScriptLanguageSchema";
 export * from "./SettingsSchema";
 export * from "./SkillSchema";
