+++ packages/schemas/src/refinement/RefinementRunSchema.ts
@@ -1,5 +1,5 @@
 import { z } from "zod/v4";
-import { MetaSchema } from "./meta";
+import { MetaSchema } from "../meta";
 
 export const RefinementRunSchema = z.object({
   id: z.string(),
